unfortunatley this wasnt feasable :cry:
there is no graphing api (within my grasp) that allows for limits (domain and range). At this point it was a waste of time for me to do this, as I already know how to use graphing librarys, and if it wasn't adding to the website either then there's really no point. I spent a lot of time researching into this but still no luck. calling this project done for now.